requirements.txt
flask==3.0.0
yt-dlp>=2026.3.17
litellm==1.30.0
python-dotenv==1.0.0
requests==2.31.0
flask-limiter==3.5.0
flask-talisman==1.1.0
